python/fast-api/requirements.txt
fastapi[all]
pydantic
psycopg2
SQLAlchemy
email-validator
passlib
bcrypt
python-jose[cryptography]
python-multipart
alembic
pydantic-settings
opentelemetry-instrumentation-fastapi